requirements.txt
slack-sdk
slack-bolt
langchain_google_genai
python-dotenv
langchain
langchain-community
Flask